test: move page-clock tests into library/ tests folder
